requirements.txt
Python==3.8
PyTorch==1.8.0
pandas==1.4.2